requirements.txt
flask==3.0.2
Pillow==10.2.0
python-dotenv==1.0.1
openai==1.12.0
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
httpx==0.27.0
authlib==1.2.0
numpy<2.0.0
h5py<3.9.0
tensorflow==2.14.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
cloudpickle>=1.6.0
deepdanbooru>=1.0.0